#793 Extracted pre action of sequential edit action in to another class.
